refactor(cli): split run.ts into directory structure for pattern consistency
Convert the monolithic run.ts (883 lines) into a directory structure
matching the established pattern used by artifact/, schedule/, volume/:

- run/index.ts: exports runCommand with subcommands wired via addCommand()
- run/run.ts: main run command (262 lines)
- run/resume.ts: resume subcommand (173 lines)
- run/continue.ts: continue subcommand (177 lines)
- run/shared.ts: shared utilities (308 lines)

Closes #1388
