+++ turbo/apps/cli/src/commands/run/index.ts
@@ -1,0 +1,9 @@
+import { mainRunCommand } from "./run";
+import { resumeCommand } from "./resume";
+import { continueCommand } from "./continue";
+
+// Add subcommands to the main run command
+mainRunCommand.addCommand(resumeCommand);
+mainRunCommand.addCommand(continueCommand);
+
+export const runCommand = mainRunCommand;
